refactor: extract DocInfoCard.vue
